+++ lib/src/HighlightedWord.dart
@@ -1,12 +1,10 @@
 import 'package:flutter/material.dart';
 
 class HighlightedWord {
-  final String meaning;
   final TextStyle textStyle;
-  final Function(BuildContext) onTap;
+  final VoidCallback onTap;
 
   HighlightedWord({
-    @required this.meaning,
     @required this.onTap,
     this.textStyle = const TextStyle(
       color: Colors.red,