+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.2.0
+The meaning attribute was removed from the HighlightedWord class; the onTap is working normally.
+
 ## 0.0.1+2
 Minor maintenance fixes
 
@@ -5,4 +8,4 @@ Minor maintenance fixes
 Example correction
 
 ## 0.0.1
-With this plugin you can highlight words and create specific actions for each highlighted word, you can customize the style of each word separately or create a unique style for all of them, you can also customize the style of the rest of the text.
+With this package you can highlight words and create specific actions for each highlighted word, you can customize the style of each word separately or create a unique style for all of them, you can also customize the style of the rest of the text.